csvw-eo-library/pyproject.toml
[project]
name = "csvw-eo"
version = "0.1.0"
description = "Safe processing utilities for CSVW data"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "DSCC, FSO", email = "dscc@example.com" }
]
urls = { Homepage = "https://github.com/dsccadminch/csvw-eo-library/csvw-eo" }
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
    "numpy>=1.21",
    "pandas>=1.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "polars==1.32.0",
    "pyarrow>=23.0.1",
    "opendp>=0.12",
    "rdflib>=6.0", 
    "pyshacl>=0.17.2"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy==1.15.0",
    "coverage==7.6.12",
    "pytest==8.3.3",
    "pytest-cov==6.0.0",
    "types-PyYAML==6.0.12.20250402",
    "markdown==3.8.2",
    "mkdocs==1.6.1"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true
exclude = "\\.ipynb_checkpoints"
disable_error_code = []

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = [
    "ANN", # strict annotations
    "ARG", # unused argument detection
    "B",   # bugbear
    "C4",  # better comprehensions
    "D",   # docstrings
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # import sorting
    "N",   # naming rules
    "PL",  # pylint like
    "RUF", # ruf opinion
    "S",   # security
    "SIM", # simplification rules
    "T20", # no print debugging
    "UP",  # pyupgrade
]
ignore = [
    "D203",
    "D212",
]

[tool.coverage.run]
branch = true
relative_files = true
source = ["csvw_eo"]        # only track your package
omit = [
    "*/tests/*",              # ignore test files
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "if __name__ == .__main__.:",   # ignore main guard
    "if TYPE_CHECKING:",            # ignore main guard
    "def main\\(\\).*:",            # ignore the main() functions
    "raise .*Error",                # ignore lines raising ValueError
]
